+++ sync-for-commerce/src/codatsynccommerce/models/operations/request_sync_for_date_range.py
@@ -8,21 +8,23 @@ from ..shared import syncsummary as shared_syncsummary
 from typing import Optional
 
 
-
 @dataclasses.dataclass
 class RequestSyncForDateRangeRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     sync_range: Optional[shared_syncrange.SyncRange] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})
     
-
 
 
 
 @dataclasses.dataclass
 class RequestSyncForDateRangeResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""
     sync_summary: Optional[shared_syncsummary.SyncSummary] = dataclasses.field(default=None)
     r"""Success"""