+++ sync-for-commerce/src/codatsynccommerce/sdk.py
@@ -9,6 +9,7 @@ from .sync import Sync
 from .sync_flow_settings import SyncFlowSettings
 from codatsynccommerce import utils
 from codatsynccommerce.models import shared
+from typing import Dict
 
 class CodatSyncCommerce:
     r"""Sync for Commerce: The API for Sync for Commerce.
@@ -36,7 +37,7 @@ class CodatSyncCommerce:
                  security: shared.Security = None,
                  server_idx: int = None,
                  server_url: str = None,
-                 url_params: dict[str, str] = None,
+                 url_params: Dict[str, str] = None,
                  client: requests_http.Session = None,
                  retry_config: utils.RetryConfig = None
                  ) -> None:
@@ -49,7 +50,7 @@ class CodatSyncCommerce:
         :param server_url: The server URL to use for all operations
         :type server_url: str
         :param url_params: Parameters to optionally template the server URL with
-        :type url_params: dict[str, str]
+        :type url_params: Dict[str, str]
         :param client: The requests.Session HTTP client to use for all operations
         :type client: requests_http.Session
         :param retry_config: The utils.RetryConfig to use globally
@@ -58,7 +59,9 @@ class CodatSyncCommerce:
         if client is None:
             client = requests_http.Session()
         
+        
         security_client = utils.configure_security_client(client, security)
+        
         
         if server_url is not None:
             if url_params is not None: