+++ sync-for-commerce/RELEASES.md
@@ -549,3 +549,13 @@ Based on:
 - [python v2.1.0] sync-for-commerce
 ### Releases
 - [PyPI v2.1.0] https://pypi.org/project/codat-sync-for-commerce/2.1.0 - sync-for-commerce
+
+## 2023-10-17 14:50:27
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.100.2 (2.159.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.0.0] sync-for-commerce
+### Releases
+- [PyPI v3.0.0] https://pypi.org/project/codat-sync-for-commerce/3.0.0 - sync-for-commerce